pyproject.toml
[tool.pdm]

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Master-Hash", email = "hash@hash.memorial"},
]
dependencies = [
    "python-docx>=0.8.11",
]
requires-python = ">=3.11"
license = {text = "Unlicense"}

[tool.pdm.scripts]
pkg.cmd = "python rename.py"
pkg.env_file = ".env"
